Space Game 0.81
Enemy sprites have been replaced with a red arrow and prototype code to make them rotate towards the player has been added. A Rectangle 'middle' has been added to the Enemy class for this rotation logic, as well as new Point enemyPos and double enemyAngle. The enemy movement logic has been fixed to move the enemy middle accordingly. Improvements are underway to make Space Game work with any monitor resolution. The form currently resizes and all dependencies on 1920x1080 monitors have been corrected to dynamically resize to the screen. Some UI elements are still non functioning. Background scroll animations have been corrected to scroll the full length of the background, while keeping the same speed. The redundant function. Redundant sections of code, such as the function GetMousePos have been removed.
