+++ 12IA Game WPF/12IA Game WPF/MainWindow.xaml
@@ -5,12 +5,12 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:_12IA_Game_WPF"
         mc:Ignorable="d"
-        Title="Space Game" Topmost="True" WindowStartupLocation="CenterScreen" ShowInTaskbar="False" Height="1080" Width="1920" ResizeMode="CanMinimize" WindowStyle="None" WindowState="Maximized">
+        Title="Space Game" Topmost="True" WindowStartupLocation="CenterScreen" ShowInTaskbar="False" Height="{x:Static SystemParameters.PrimaryScreenHeight}" Width="{x:Static SystemParameters.PrimaryScreenHeight}" ResizeMode="CanMinimize" WindowStyle="None" WindowState="Maximized">
 
     <Canvas Background="LightBlue" Name="Game_Canvas" Focusable="True" KeyDown="Game_Canvas_KeyDown" KeyUp="Game_Canvas_KeyUp">
 
         <Image x:Name="imgBackground" Height="1080" Width="3840" Source="/images/space background 2.png"/>
-        <Button x:Name="btnExit" Content="runurunrunrunr" Canvas.Left="890" Canvas.Top="755" Width="140" Height="56" Click="Exit"/>
+        <Button x:Name="btnExit" Content="runurunrunrunr" Canvas.Left="470" Canvas.Top="729" Width="140" Height="56" Click="Exit"/>
 
     </Canvas>
 </Window>
